config: Move input config handling into seperate file
